moved the code for vertex-disjoint-paths from the algorithm for directed-tree-decompositions to its own algorithm. also prepared an algorithm for vertex-separators which will use the vertex-disjoint-paths problem (see mengers theorem) and referenced that vertex-separator algorithm in the header for directed-tree-decompositions.
further, the method RunParallel(...) of the "Algorithm" base-class
now returns the result again (at least in case there is exactly 1
algorithm applicable)

+++ opengt.so/Sources/opengt.cpp
@@ -1302,7 +1302,7 @@ namespace OpenGraphtheory
             fattr->Value = weight;
         }
 
-        Coordinates GraphObject::GetCoordinates(int minCount, string name)
+        Coordinates GraphObject::GetCoordinates(size_t minCount, string name)
         {
             Coordinates result;
             SeqAttribute* sattr = dynamic_cast<SeqAttribute*>(GetAttribute(name));
